LPS-145947 Creates a new *not public* editor component called BaseEditor and use it
* Creates a new React component for being a base for other Editor components: BaseEditor

Now, with this modification we have the following hierarchy:

```
	ClassicEditor / BalloonEditor

			BaseEditor

				Editor
```

With this change, we don't need to deprecate Editor component and we can avoid duplication inside ClassicEditor/BalloonEditor
