+++ modules/apps/frontend-editor/frontend-editor-ckeditor-web/src/main/resources/META-INF/resources/editor/ClassicEditor.js
@@ -13,74 +13,82 @@
  */
 
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, {forwardRef} from 'react';
 
-import {Editor} from './Editor';
+import BaseEditor from './BaseEditor';
 
-const ClassicEditor = ({
-	contents,
-	editorConfig,
-	initialToolbarSet = 'simple',
-	name,
-	title,
-	...otherProps
-}) => {
-	return (
-		<div id={`${name}Container`}>
-			{title && (
-				<label className="control-label" htmlFor={name}>
-					{title}
-				</label>
-			)}
+const ClassicEditor = forwardRef(
+	(
+		{
+			contents,
+			editorConfig,
+			initialToolbarSet = 'simple',
+			name,
+			title,
+			...otherProps
+		},
+		ref
+	) => {
+		return (
+			<div id={`${name}Container`}>
+				{title && (
+					<label className="control-label" htmlFor={name}>
+						{title}
+					</label>
+				)}
 
-			<Editor
-				className="lfr-editable"
-				config={{
-					toolbar: initialToolbarSet,
-					...editorConfig,
-				}}
-				contents={contents}
-				name={name}
-				onBeforeLoad={(CKEDITOR) => {
-					CKEDITOR.disableAutoInline = true;
-					CKEDITOR.dtd.$removeEmpty.i = 0;
-					CKEDITOR.dtd.$removeEmpty.span = 0;
+				<BaseEditor
+					className="lfr-editable"
+					config={{
+						toolbar: initialToolbarSet,
+						...editorConfig,
+					}}
+					contents={contents}
+					name={name}
+					onBeforeLoad={(CKEDITOR) => {
+						CKEDITOR.disableAutoInline = true;
+						CKEDITOR.dtd.$removeEmpty.i = 0;
+						CKEDITOR.dtd.$removeEmpty.span = 0;
 
-					CKEDITOR.getNextZIndex = function () {
-						return CKEDITOR.dialog._.currentZIndex
-							? CKEDITOR.dialog._.currentZIndex + 10
-							: Liferay.zIndex.WINDOW + 10;
-					};
-				}}
-				onDrop={(event) => {
-					const data = event.data.dataTransfer.getData('text/html');
-					const editor = event.editor;
-
-					if (data) {
-						const fragment = CKEDITOR.htmlParser.fragment.fromHtml(
-							data
+						CKEDITOR.getNextZIndex = function () {
+							return CKEDITOR.dialog._.currentZIndex
+								? CKEDITOR.dialog._.currentZIndex + 10
+								: Liferay.zIndex.WINDOW + 10;
+						};
+					}}
+					onDrop={(event) => {
+						const data = event.data.dataTransfer.getData(
+							'text/html'
 						);
+						const editor = event.editor;
 
-						const name = fragment.children[0].name;
+						if (data) {
+							const fragment = CKEDITOR.htmlParser.fragment.fromHtml(
+								data
+							);
 
-						if (name) {
-							return editor.pasteFilter.check(name);
+							const name = fragment.children[0].name;
+
+							if (name) {
+								return editor.pasteFilter.check(name);
+							}
 						}
-					}
-				}}
-				onInstanceReady={({editor}) => {
-					editor.setData(contents, {
-						callback: () => {
-							editor.resetUndo();
-						},
-						noSnapshot: true,
-					});
-				}}
-				{...otherProps}
-			/>
-		</div>
-	);
-};
+					}}
+					onInstanceReady={({editor}) => {
+						editor.setData(contents, {
+							callback: () => {
+								editor.resetUndo();
+							},
+							noSnapshot: true,
+						});
+					}}
+					ref={ref}
+					{...otherProps}
+				/>
+			</div>
+		);
+	}
+);
 
 ClassicEditor.propTypes = {
 	contents: PropTypes.string,
